+++ fdbserver/Knobs.h
@@ -381,6 +381,7 @@ public:
 	int64_t TIME_KEEPER_DELAY;
 	int64_t TIME_KEEPER_MAX_ENTRIES;
 
+
 	ServerKnobs(bool randomize = false, ClientKnobs* clientKnobs = NULL);
 };